requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
easyocr>=1.7.0
python-docx>=0.8.11
openai>=1.3.7
python-dotenv>=1.0.0